Update styles.css
Added a new grid class for 5-up display, “fifths”.
Renamed the existing “fifths” class to “forty-sixty” to represent the 40/60 content split.
Added ability to apply “dark” class to en entire block of icons, in addition to individual icons. This also requires the entire icon group be wrapped in “icon-container.”
Added a “smaller” class for reducing icon size in smaller grids. This also requires the entire icon group be wrapped in “icon-container.”